+++ src/main/java/com/gendev/streamcombinations/model/StreamingOffer.java
@@ -11,7 +11,7 @@ import jakarta.persistence.Id;
 public class StreamingOffer {
 
     @Id
-    private Long id;
+    private int id;
 
     private int streaming_package_id;
     private int live;
@@ -20,7 +20,7 @@ public class StreamingOffer {
     public StreamingOffer() {
     }
 
-    public StreamingOffer(Long id, int streaming_package_id, int live, int highlights) {
+    public StreamingOffer(int id, int streaming_package_id, int live, int highlights) {
         this.id = id;
         this.streaming_package_id = streaming_package_id;
         this.live = live;
